关闭计划已加入仍可学

+++ EduSohoAndroidClient/edusoho_v3/src/main/java/com/edusoho/kuozhi/clean/bean/CourseLearningProgress.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class CourseLearningProgress implements Serializable {
 
     public int taskCount;
-    public int progress;
+    public float progress;
     public int taskResultCount;
     public int taskPerDay;
     public int planStudyTaskCount;
